+++ Sources/ObjC/include/Honeybadger.h
@@ -16,14 +16,8 @@ NS_ASSUME_NONNULL_BEGIN
 + (void) configureWithAPIKey:(NSString*)apiKey
 NS_SWIFT_NAME(configure(apiKey:));
 
-+ (void) configureWithAPIKey:(NSString*)apiKey fingerprint:(NSString*)fingerprint
-NS_SWIFT_NAME(configure(apiKey:fingerprint:));
-
 + (void) configureWithAPIKey:(NSString*)apiKey environment:(NSString*)environment
 NS_SWIFT_NAME(configure(apiKey:environment:));
-
-+ (void) configureWithAPIKey:(NSString*)apiKey environment:(NSString*)environment fingerprint:(NSString*)fingerprint
-NS_SWIFT_NAME(configure(apiKey:environment:fingerprint:));
 
 // NOTIFY ------------------------------------------------------------------
 
@@ -36,8 +30,22 @@ NS_SWIFT_NAME(notify(errorString:errorClass:));
 + (void) notifyWithString:(NSString*)errorString context:(NSDictionary<NSString*, NSString*>*)context
 NS_SWIFT_NAME(notify(errorString:context:));
 
++ (void) notifyWithString:(NSString*)errorString fingerprint:(NSString*)fingerprint
+NS_SWIFT_NAME(notify(errorString:fingerprint:));
+
 + (void) notifyWithString:(NSString*)errorString errorClass:(NSString*)errorClass context:(NSDictionary<NSString*, NSString*>*)context
 NS_SWIFT_NAME(notify(errorString:errorClass:context:));
+
++ (void) notifyWithString:(NSString*)errorString errorClass:(NSString*)errorClass fingerprint:(NSString*)fingerprint
+NS_SWIFT_NAME(notify(errorString:errorClass:fingerprint:));
+
++ (void) notifyWithString:(NSString*)errorString context:(NSDictionary<NSString*, NSString*>*)context fingerprint:(NSString*)fingerprint
+NS_SWIFT_NAME(notify(errorString:context:fingerprint:));
+
++ (void) notifyWithString:(NSString*)errorString errorClass:(NSString*)errorClass context:(NSDictionary<NSString*, NSString*>*)context fingerprint:(NSString*)fingerprint
+NS_SWIFT_NAME(notify(errorString:errorClass:context:fingerprint:));
+
+// ---
 
 + (void) notifyWithError:(NSError*)error
 NS_SWIFT_NAME(notify(error:));
@@ -45,8 +53,23 @@ NS_SWIFT_NAME(notify(error:));
 + (void) notifyWithError:(NSError*)error errorClass:(NSString*)errorClass
 NS_SWIFT_NAME(notify(error:errorClass:));
 
++ (void) notifyWithError:(NSError*)error context:(NSDictionary<NSString*, NSString*>*)context
+NS_SWIFT_NAME(notify(error:context:));
+
++ (void) notifyWithError:(NSError*)error fingerprint:(NSString*)fingerprint
+NS_SWIFT_NAME(notify(error:fingerprint:));
+
 + (void) notifyWithError:(NSError*)error errorClass:(NSString*)errorClass context:(NSDictionary<NSString*, NSString*>*)context
 NS_SWIFT_NAME(notify(error:errorClass:context:));
+
++ (void) notifyWithError:(NSError*)error errorClass:(NSString*)errorClass fingerprint:(NSString*)fingerprint
+NS_SWIFT_NAME(notify(error:errorClass:fingerprint:));
+
++ (void) notifyWithError:(NSError*)error context:(NSDictionary<NSString*, NSString*>*)context fingerprint:(NSString*)fingerprint
+NS_SWIFT_NAME(notify(error:context:fingerprint:));
+
++ (void) notifyWithError:(NSError*)error errorClass:(NSString*)errorClass context:(NSDictionary<NSString*, NSString*>*)context fingerprint:(NSString*)fingerprint
+NS_SWIFT_NAME(notify(error:errorClass:context:fingerprint:));
 
 // CONTEXT -----------------------------------------------------------------